// apps/frontend/src/api/generated/model/goalOut.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * WDIAG — Where Did It All Go
 * Personal finance budgeting and intelligence
 * OpenAPI spec version: 0.2.32
 */
import type { GoalOutDescription } from './goalOutDescription'
import type { GoalType } from './goalType'
import type { GoalStatus } from './goalStatus'
import type { GoalOutTargetAmount } from './goalOutTargetAmount'
import type { GoalOutTargetDate } from './goalOutTargetDate'
import type { FundingStrategy } from './fundingStrategy'
import type { CompletionPolicy } from './completionPolicy'
import type { GoalOutAutoExtendAmount } from './goalOutAutoExtendAmount'
import type { GoalOutOwnerId } from './goalOutOwnerId'
import type { GoalOutLinkedDebtPlanId } from './goalOutLinkedDebtPlanId'
import type { GoalOutLinkedCategoryId } from './goalOutLinkedCategoryId'
import type { GoalOutMinimumBalanceThreshold } from './goalOutMinimumBalanceThreshold'
import type { GoalOutMetadata } from './goalOutMetadata'

export interface GoalOut {
  id: string
  household_id: string
  name: string
  description: GoalOutDescription
  goal_type: GoalType
  status: GoalStatus
  target_amount: GoalOutTargetAmount
  currency: string
  target_date: GoalOutTargetDate
  funding_strategy: FundingStrategy
  completion_policy: CompletionPolicy
  auto_extend_amount: GoalOutAutoExtendAmount
  owner_id: GoalOutOwnerId
  linked_debt_plan_id: GoalOutLinkedDebtPlanId
  linked_category_id: GoalOutLinkedCategoryId
  minimum_balance_threshold: GoalOutMinimumBalanceThreshold
  metadata_: GoalOutMetadata
  created_at: string
  updated_at: string
}